requirements.txt
amqp==5.0.7
asgiref==3.4.1
billiard==3.6.4.0
celery==5.2.1
cffi==1.15.0
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
cryptography==35.0.0
Deprecated==1.2.13
Django==3.2.9
django-rest-knox==4.1.0
djangorestframework==3.12.4
kombu==5.2.2
Pillow==8.4.0
prompt-toolkit==3.0.24
pycparser==2.21
pytz==2021.3
redis==4.0.2
six==1.16.0
sqlparse==0.4.2
vine==5.0.0
wcwidth==0.2.5
wrapt==1.13.3
